+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - support for `body-tracking` modality
 - support for multimodality for XRoom dataset (body tracking + bio-measurements) modalities.
+- training benchmarks for body-tracking and bio-measurements
 
 ### Changed 
 - update services names